feat: install VueRouter
VueRouterを追加
componentのimportに@を使わないとエラーが出る模様
