+++ package.json
@@ -14,6 +14,7 @@
     "@fortawesome/vue-fontawesome": "^0.1.7",
     "core-js": "^2.6.5",
     "vue": "^2.6.10",
+    "vue-router": "^3.1.3",
     "vuetify": "^2.0.0"
   },
   "devDependencies": {
