+++ src/components/Home.vue
@@ -1,0 +1,15 @@
+<template>
+<v-app>
+    <v-app-bar app>
+      <v-toolbar-title class="headline text-uppercase">
+        <span>ienage.com</span>
+        <!-- <span class="font-weight-light">COUNT UP</span> -->
+      </v-toolbar-title>
+    </v-app-bar>
+    <v-content>
+    <h1><router-link to="/countup">COUNTUP</router-link></h1>
+    <h1>SHOOTOUT(近日公開)</h1>
+    </v-content>
+  </v-app>
+
+</template>
