+++ src/components/ShootOut.vue
@@ -1,0 +1,17 @@
+<template>
+  <v-app>
+    <v-app-bar app>
+      
+    </v-app-bar>
+  </v-app>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+
+</style>
